# docs/api/swagger.yaml
basePath: /
definitions:
  github_com_zhian9_blogo-server_pkg_errors.Error:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      detail:
        description: 详细描述
        type: string
      id:
        description: 错误标识（前端友好）
        type: string
      status:
        description: 状态文本（如 "Not Found"）
        type: string
    type: object
...
  schema.LoginForm:
    properties:
      captcha_code:
        description: |-
          CaptchaCode: 用户输入的验证码（必填）
          前端从验证码图片中识别的字符
          Example: aB3x9
        example: aB3x9
        type: string
      captcha_id:
        description: |-
          CaptchaID: 验证码 ID（必填）
          来自 Captcha 接口的 captcha_id
          Example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      password:
        description: |-
          Password: 登录密码（必填）
          Example: goadmin123456
        example: goadmin123456
        type: string
      username:
        description: |-
          Username: 登录用户名（必填）
          Example: admin
        example: admin
        type: string
    required:
    - captcha_code
    - captcha_id
    - password
    - username
    type: object
  schema.LoginToken:
    properties:
      access_token:
        description: |-
          AccessToken: 访问令牌（JWT 字符串）
          前端需在后续请求的 Authorization 头中携带
          Example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.xxxxx
        example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.xxxxx
        type: string
      expires_at:
        description: |-
          ExpiresAt: 令牌过期时间（Unix 时间戳，秒）
          前端可用于自动刷新令牌
          Example: 1730466240
        example: 1730466240
        type: integer
      token_type:
        description: |-
          TokenType: 令牌类型（通常为 "Bearer"）
          使用方式: Authorization: ${token_type} ${access_token}
          Example: Bearer
        example: Bearer
        type: string
    type: object
  schema.Menu:
    properties:
      children:
        description: 子菜单列表（用于前端树形展示）
        items:
          $ref: '#/definitions/schema.Menu'
        type: array
      code:
        description: 菜单编码（同级唯一，用于权限标识）
        type: string
      created_at:
        description: 创建时间
        type: string
      description:
        description: 描述信息
        type: string
      id:
        description: 唯一 ID（主键）
        type: string
      name:
        description: 显示名称
        type: string
      parent_id:
        description: 父菜单 ID（根菜单为空）
        type: string
      parent_path:
        description: 父菜单路径（如 "root.parent"，用于快速查询祖先）
        type: string
      path:
        description: 访问路径（前端路由或 API 路径）
        type: string
      properties:
        description: '扩展属性（JSON 格式，如 {"icon": "user", "hidden": true}）'
        type: string
      resources:
        description: 关联的资源列表（API 路径+方法，用于 Casbin 权限）
        items:
          $ref: '#/definitions/schema.MenuResource'
        type: array
      sequence:
        description: 排序序号（越大越靠前）
        type: integer
      status:
        description: 状态（enabled/disabled）
        type: string
      type:
        description: 菜单类型（page/button）
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  schema.MenuForm:
    properties:
      code:
        description: 菜单编码（必填，同级唯一）
        maxLength: 32
        type: string
      description:
        description: 描述信息
        type: string
      name:
        description: 显示名称（必填）
        maxLength: 128
        type: string
      parent_id:
        description: 父菜单 ID
        type: string
      path:
        description: 访问路径
        type: string
      properties:
        description: 扩展属性（JSON 格式）
        type: string
      resources:
        description: 关联的资源列表
        items:
          $ref: '#/definitions/schema.MenuResource'
        type: array
      sequence:
        description: 排序序号
        type: integer
      status:
        description: 状态（必填，enabled/disabled）
        enum:
        - disabled
        - enabled
        type: string
      type:
        description: 菜单类型（必填，page/button）
        enum:
        - page
        - button
        type: string
    required:
    - code
    - name
    - status
    - type
    type: object
  schema.MenuResource:
    properties:
      created_at:
        description: 创建时间
        type: string
      id:
        description: 唯一 ID（主键）
        type: string
      menu_id:
        description: 关联的菜单 ID（外键，指向 menu.id）
        type: string
      method:
        description: HTTP 方法（如 GET, POST, PUT, DELETE）
        type: string
      path:
        description: |-
          API 请求路径（支持路径参数）
          示例: /api/v1/users, /api/v1/users/:id
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  schema.RegisterForm:
    properties:
      captcha_code:
        description: |-
          CaptchaCode: 验证码输入（必填）
          Example: xK9m2
        example: xK9m2
        type: string
      captcha_id:
        description: |-
          CaptchaID: 验证码 ID（必填）
          Example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        type: string
      confirm_password:
        description: |-
          ConfirmPassword: 二次确认密码（必填）
          Example: MyPass123!
        example: MyPass123!
        maxLength: 64
        type: string
      email:
        description: |-
          Email: 邮箱（必填，合法格式）
          Example: user2025@example.com
        example: user2025@example.com
        maxLength: 128
        type: string
      password:
        description: |-
          Password: 登录密码（必填）
          Example: MyPass123!
        example: MyPass123!
        maxLength: 64
        type: string
      phone:
        description: |-
          Phone: 手机号（必填）
          Example: +8613900001111
        example: "+8613900001111"
        maxLength: 32
        type: string
      username:
        description: |-
          Username: 登录用户名（必填）
          Example: newuser2025
        example: newuser2025
        maxLength: 64
        type: string
    required:
    - captcha_code
    - captcha_id
    - confirm_password
    - email
    - password
    - phone
    - username
    type: object
...
  schema.UpdateCurrentUser:
    properties:
      avatar:
        description: |-
          Avatar: 头像 URL（可选，最大 512 字符）
          Example: https://example.com/avatar.jpg
        example: https://example.com/avatar.jpg
        maxLength: 512
        type: string
      bio:
        description: |-
          Bio: 个人简介/个性签名（可选，最大 512 字符）
          Example: 全栈开发者，热爱开源
        example: 全栈开发者，热爱开源
        maxLength: 512
        type: string
      email:
        description: |-
          Email: 邮箱地址（可选，最大 128 字符）
          Example: lxy911@gmail.com
        example: lxy911@gmail.com
        maxLength: 128
        type: string
      name:
        description: |-
          Name: 用户姓名（必填，最大 64 字符）
          Example: 李星云
        example: 李星云
        maxLength: 64
        type: string
      phone:
        description: |-
          Phone: 手机号码（可选，最大 32 字符）
          Example: +8613800138000
        example: "+8613800138000"
        maxLength: 32
        type: string
      remark:
        description: |-
          Remark: 备注信息（可选，最大 1024 字符）
          Example: 开发者
        example: 开发者
        maxLength: 1024
        type: string
    required:
    - name
    type: object
  schema.UpdateLoginPassword:
    properties:
      new_password:
        description: |-
          NewPassword: 新密码
          Example: newSecurePass!2025
        example: newSecurePass!2025
        type: string
      old_password:
        description: |-
          OldPassword: 旧密码
          用于验证用户身份
          Example: oldPass123
        example: oldPass123
        type: string
    required:
    - new_password
    - old_password
    type: object
  schema.User:
    properties:
      activated_at:
        description: Account activation time
        type: string
      avatar:
        description: Avatar URL
        type: string
      bio:
        description: Bio / personal signature
        type: string
      created_at:
        description: Create time
        type: string
      email:
        description: Email of user (unique)
        type: string
      follower_count:
        description: 粉丝数
        type: integer
      following_count:
        description: 关注数
        type: integer
      id:
        description: Unique ID
        type: string
      last_login_at:
        description: Last login time
        type: string
      last_login_ip:
        description: Last login IP (supports IPv6)
        type: string
      name:
        description: Name of user
        type: string
      phone:
        description: Phone number of user
        type: string
      remark:
        description: Remark of user
        type: string
      roles:
        description: Roles of user
        items:
          $ref: '#/definitions/schema.UserRole'
        type: array
      status:
        description: Status of user (activated, inactive, freezed)
        type: string
      updated_at:
        description: Update time
        type: string
      username:
        description: Username for login
        type: string
    type: object
...
  schema.UserRole:
    properties:
      created_at:
        description: Create time
        type: string
      id:
        description: Unique ID
        type: string
      role_code:
        description: From Role.Code
        type: string
      role_id:
        description: From Role.ID
        type: string
      role_name:
        description: From Role.Name
        type: string
      updated_at:
        description: Update time
        type: string
      user_id:
        description: From User.ID
        type: string
    type: object
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_zhian9_blogo-server_pkg_errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
host: localhost:8040
info:
  contact:
    email: lxyaa911@gmail.com
    name: 李星云 (lxy911)
    url: https://github.com/zhian9
  description: Modern SaaS Blog Platform — REST API for content management, user system,
    RBAC, analytics, and more.
  license:
    name: MIT
    url: https://github.com/zhian9/Blogo/blob/master/LICENSE
  termsOfService: https://github.com/zhian9/Blogo
  title: Blogo API
  version: v1.0.0
paths:
...
  /api/v1/captcha/image:
    get:
      parameters:
      - description: Captcha ID
        in: query
        name: id
        required: true
        type: string
      - description: Reload captcha image (reload=1)
        in: query
        name: reload
        type: number
      produces:
      - image/png
      responses:
        "200":
          description: Captcha image
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Response captcha image
      tags:
      - LoginAPI
...
  /api/v1/current/logout:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Logout system
      tags:
      - LoginAPI
  /api/v1/current/menus:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Query current user menus based on the current user role
      tags:
      - LoginAPI
  /api/v1/current/password:
    put:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateLoginPassword'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Change current user password
      tags:
      - LoginAPI
  /api/v1/current/refresh-token:
    post:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.LoginToken'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Refresh current access token
      tags:
      - LoginAPI
  /api/v1/current/user:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - LoginAPI
    put:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateCurrentUser'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Update current user info
      tags:
      - LoginAPI
...
  /api/v1/login:
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.LoginForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.LoginToken'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Login system with username and password
      tags:
      - LoginAPI
  /api/v1/menus:
    get:
      parameters:
      - description: Code path of menu (like xxx.xxx.xxx)
        in: query
        name: code
        type: string
      - description: Name of menu
        in: query
        name: name
        type: string
      - description: Whether to include menu resources
        in: query
        name: includeResources
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Menu'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Query menu tree data
      tags:
      - MenuAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.MenuForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Create menu record
      tags:
      - MenuAPI
  /api/v1/menus/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Delete menu record by ID
      tags:
      - MenuAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Menu'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Get menu record by ID
      tags:
      - MenuAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.MenuForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - BearerAuth: []
      summary: Update menu record by ID
      tags:
      - MenuAPI
...
  /api/v1/register:
    post:
      description: Register a new user account. An activation email will be sent to
        the provided email.
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.RegisterForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Register a new user
      tags:
      - LoginAPI
...
  /api/v1/verify-email:
    get:
      description: Activate user account by clicking the activation link sent via
        email.
      parameters:
      - description: Activation token from email
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      summary: Verify email via activation token
      tags:
      - LoginAPI
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your JWT token>". Login via /api/v1/login to obtain
      a token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"